refactor: Move EPC DTOs to separate file
